Fixed a bug causing sigsegv for a WMList with more than 32767 items

+++ WINGs/wlist.c
@@ -16,7 +16,7 @@ typedef struct W_List {
 
     short itemHeight;
 
-    short topItem;	    /* index of first visible item */
+    int topItem;	    /* index of first visible item */
 
     short fullFitLines;	    /* no of lines that fit entirely */
